pyproject.toml
[project]
name = "ship-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.129.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
